requirements-dev.txt
-r requirements.txt

# General
# ------------------------------------------------------------------------------
Werkzeug[watchdog]==2.3.6
ipdb==0.13.13
watchfiles==0.19.0

# Code quality
# ------------------------------------------------------------------------------
flake8==6.0.0
flake8-isort==6.0.0
coverage==7.2.7
black==23.7.0
djlint==1.32.1
pylint-django==2.5.3
pre-commit==3.3.3
